+++ AssessmentDatabase/ValueObjects/EntitiesFields.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 public protocol Sidable {
-    var sid: Int { get set }
+    var sid: Int64 { get set }
 }
 
 public protocol InstructorFields: Sidable {
@@ -18,16 +18,16 @@ public protocol InstructorFields: Sidable {
 }
 
 public protocol StudentFields: Sidable {
-    var assessmentSids: [Int] { get set }
-    var instructorSids: [Int] { get set }
-    var microTaskGradesSids: [Int] { get set }
+    var assessmentSids: [Int64] { get set }
+    var instructorSids: [Int64] { get set }
+    var microTaskGradesSids: [Int64] { get set }
 }
 
 public protocol AssessmentFields: Sidable {
     var date: Date { get set }
-    var schoolId: Int { get set }
+    var schoolId: Int64 { get set }
     
-    var instructorSid: Int { get set }
+    var instructorSid: Int64 { get set }
     var rubric: RubricFields { get set }
     var studentMicrotaskGrades: [StudentMicrotaskGradeFields] { get set }
     var students: [StudentFields] { get set }
@@ -38,19 +38,19 @@ public protocol RubricFields: Sidable {
 }
 
 public protocol SkillSetFields: Sidable {
-    var rubricSid: Int { get set }
+    var rubricSid: Int64 { get set }
     var microTasks: [MicrotaskFields] { get set }
 }
 
 public protocol MicrotaskFields: Sidable {
-    var skillSetSid: Int { get set }
+    var skillSetSid: Int64 { get set }
 }
 
 public protocol StudentMicrotaskGradeFields: Sidable {
-    var assessmentSid: Int { get set }
+    var assessmentSid: Int64 { get set }
     var grade: GradeFields { get set }
-    var microTaskSid: Int { get set }
-    var studentSid: Int { get set }
+    var microTaskSid: Int64 { get set }
+    var studentSid: Int64 { get set }
 }
 
 public protocol GradeFields: Sidable {
